fixes to test suite. pushing to run screenshot tests
